style: Update prettier, eslint, stylelint config. Switch to single quotes and require semicolons in JS code. Run stylelint and prettier to fix style

+++ stylelint.config.js
@@ -3,7 +3,7 @@ const stylelintConfig = require('@nextcloud/stylelint-config')
 stylelintConfig.extends.push('stylelint-config-idiomatic-order')
 
 stylelintConfig.rules.indentation = null
-stylelintConfig.rules['string-quotes'] = 'double'
+stylelintConfig.rules['string-quotes'] = 'single'
 stylelintConfig.rules['function-no-unknown'] = [true, {
 	'ignoreFunctions': ["math.div"]
 }]
